Projeto Hugging/requirements.txt
discord.py
anthropic
requests